Switch to ruby 4.0 as a default ruby version for CI (#25)
This pull request updates the CI workflows to use Ruby 4.0 as the
primary version instead of Ruby 3.4, and adds support for testing
against the latest Ruby version (`ruby-head`). The changes ensure that
tests and quality checks are run on the latest Ruby versions, with
special handling to allow failures on the bleeding-edge `ruby-head`
build.